[Serialization] Serialize and check the target a module was compiled with.
Refuse to load a module if it was compiled for a different architecture or
OS, or if its minimum deployment target is newer than the current target.
Additionally, provide the target triple as part of pre-loading validation
for clients who care (like LLDB).

Part of rdar://problem/17670778

Swift SVN r24469

+++ test/IRGen/class_with_stub_initializers.swift
@@ -1,6 +1,6 @@
 // RUN: rm -rf %t && mkdir %t
 // RUN: %build-irgen-test-overlays
-// RUN: %swift -target x86_64-apple-macosx10.9 -sdk %S/Inputs -I %t %s -emit-ir | FileCheck %s
+// RUN: %swift -sdk %S/Inputs -I %t %s -emit-ir | FileCheck %s
 
 import Foundation
 
@@ -12,5 +12,5 @@ import Foundation
 // CHECK-NOT:     %[[FOO]]* (%[[FOO]]*)*
 // CHECK:       {{^(@|define)}}
 class Foo: NSObject {
-  init(x: Int) { super.init() }
+  init(x: Int64) { super.init() }
 }
